Fix naccept type: float -> size_t in bh_step_data
naccept is a step counter, not a fractional value. Declared float
was inconsistent with nstep (size_t) and risked precision loss for
large run counts. Add explicit (double) cast in accept_rate division
to preserve correct floating-point result.

+++ cfl/include/basinhopping.h
@@ -37,7 +37,7 @@ typedef struct {
   /* Number of steps taken. */
   size_t nstep;
   /* Number of accepted steps. */
-  float naccept;
+  size_t naccept;
   /* Target acceptance rate. */
   float target_accept_rate;
   /* Interval for how often to update stepsize. */
